Revert test config change
