--- tests/configs/utils.py
@@ -1,10 +1,0 @@
-import os
-
-
-def get_all_toml_files(directory):
-    toml_files = []
-    for root, _, files in os.walk(directory):
-        for file in files:
-            if file.endswith(".toml"):
-                toml_files.append(os.path.join(root, file))
-    return toml_files
